Display posts in DOM, file reorganization, adjusted api url hook
App-blog.js now has functionality to render all posts in the DOM
with bootstrapped divs.
Reorganized files to place appropriate modules in public folder
and move other modules to private.
Finally, adjusted the url webhook for the api to /api/posts
rather than /posts in server.js and in DATABASE_URL in app-blog.
This seems to be best practice rather than just having the api
hook on to the base url without an additional prefix.